# Conflicts:
#	geoapps/block_model_creation/notebook.ipynb
#	geoapps/index.ipynb

+++ geoapps/block_model_creation/params.py
@@ -117,7 +117,7 @@ class BlockModelParams(BaseParams):
         self.setter_validator("bottom_padding", val)
 
     @property
-    def depth_core(self) -> int | None:
+    def depth_core(self) -> float | None:
         """
         Core depth for 3D grid.
         """
